Extracted all output messages to a Reporter class
Supporting a usecase for silencing output messages. A ‘:shush’ options
can be set to true to enable a silent mode. By default :shush is nil
and so the silent mode won’t be turned on for existing usecases, unless
explicitly set to true in options. stdout sync is set to true to avoid
buffering.